requirements.txt
Django==4.2
djangorestframework==3.14.0
drf-yasg==1.21.4
celery==5.3.0
redis==4.5.4